Extract highlight decorations to a separate file (#168185)
* Extract highlight decorations to a separate file

* Fix CSS import
