+++ src/vs/editor/contrib/wordHighlighter/browser/highlightDecorations.css
